update sast-java cases
