+++ sast-java/src/main/java/com/sast/astbenchmark/model/custom/ABS.java
@@ -1,0 +1,9 @@
+package com.sast.astbenchmark.model.custom;
+
+public abstract class ABS {
+    protected String cmd;
+    public void setCmd(String cmd){
+        this.cmd = cmd;
+    }
+    public abstract String getCmd();
+}
