+++ sast-java/src/main/java/com/sast/astbenchmark/model/custom/AC.java
@@ -1,0 +1,8 @@
+package com.sast.astbenchmark.model.custom;
+
+public class AC extends ABS{
+    @Override
+    public String getCmd(){
+        return super.cmd;
+    }
+}
