+++ sast-java/src/main/java/com/sast/astbenchmark/model/custom/O.java
@@ -1,0 +1,20 @@
+package com.sast.astbenchmark.model.custom;
+
+public class O {
+    private String cmd;
+
+    public O(String cmd){
+        this.cmd = cmd;
+    }
+
+    class I {
+        String getCmd(){
+            return O.this.cmd;
+        }
+    }
+
+    public String getCmd(){
+        O.I i = new I();
+        return i.getCmd();
+    }
+}
