+++ sast-java/src/main/java/com/sast/astbenchmark/model/custom/P.java
@@ -1,0 +1,15 @@
+package com.sast.astbenchmark.model.custom;
+
+public class P {
+    private String cmd;
+
+    public P(String cmd) {
+        this.cmd = cmd;
+    }
+
+    public String getCmd() {
+        return this.cmd;
+    }
+
+}
+
