+++ sast-java/src/main/java/com/sast/astbenchmark/model/custom/PS.java
@@ -1,0 +1,19 @@
+package com.sast.astbenchmark.model.custom;
+
+public class PS extends P {
+    private String cmd2;
+
+    public PS(String cmd, String cmd2) {
+        super(cmd);
+        this.cmd2 = cmd2;
+    }
+
+    public String getCmd() {
+        return super.getCmd() + this.cmd2;
+    }
+
+    public String getCmd(String cmd) {
+        return cmd + this.cmd2;
+    }
+
+}
